requirements.txt
selenium~=3.141.0
PyYAML~=5.3.1